submission/Model Deployment & Inference/Docker/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
albumentations
opencv-python-headless
pillow
numpy
scikit-learn
torch
